Put related things together (in-memory store, errors, traits, types) and don't use the internal version types of the in-memory store in public APIs.
