+++ src/zonefile/mod.rs
@@ -2,5 +2,7 @@
 #![cfg(feature = "zonefile")]
 #![cfg_attr(docsrs, doc(cfg(feature = "zonefile")))]
 
+pub mod error;
 pub mod inplace;
+#[cfg(feature = "unstable-zonetree")]
 pub mod parsed;
