+++ src/zonefile/parsed.rs
@@ -1,21 +1,21 @@
 //! Importing from and exporting to a zonefiles.
 
-use tracing::trace;
-
-use crate::base::iana::{Class, Rtype};
-use crate::base::name::FlattenInto;
-use crate::base::ToDname;
-use crate::rdata::ZoneRecordData;
-use crate::zonetree::{
-    CnameError, Rrset, SharedRr, StoredDname, StoredRecord, ZoneBuilder,
-    ZoneCutError,
-};
 use core::convert::Infallible;
 use std::collections::{BTreeMap, HashMap};
 use std::fmt::Display;
 use std::vec::Vec;
 
+use tracing::trace;
+
+use super::error::{CnameError, ZoneCutError};
 use super::inplace::{self, Entry};
+
+use crate::base::iana::{Class, Rtype};
+use crate::base::name::FlattenInto;
+use crate::base::ToDname;
+use crate::rdata::ZoneRecordData;
+use crate::zonetree::in_memory::ZoneBuilder;
+use crate::zonetree::{Rrset, SharedRr, StoredDname, StoredRecord};
 
 //------------ Zonefile ------------------------------------------------------
