+++ src/zonetree/in_memory/mod.rs
@@ -1,0 +1,7 @@
+mod builder;
+mod nodes;
+mod read;
+mod versioned;
+mod write;
+
+pub use builder::ZoneBuilder;
