+++ src/zonetree/mod.rs
@@ -1,23 +1,25 @@
-#![cfg(feature = "zonefile")]
-#![cfg_attr(docsrs, doc(cfg(feature = "zonefile")))]
+//! Building, querying and iterating over zone trees.
+//!
+//!
+#![cfg(feature = "unstable-zonetree")]
+#![cfg_attr(docsrs, doc(cfg(feature = "unstable-zonetree")))]
 
-pub use self::builder::{CnameError, ZoneBuilder, ZoneCutError};
-pub use self::nodes::OutOfZone;
-pub(crate) use self::read::ReadZone;
-pub use self::read::{Answer, AnswerContent, ReadableZone, WalkOp};
-pub use self::rrset::{
+mod tree;
+mod types;
+mod walk;
+mod zone;
+
+pub mod answer;
+pub mod in_memory;
+pub mod traits;
+
+pub use self::answer::Answer;
+pub use self::traits::{
+    ReadableZone, WriteableZone, WriteableZoneNode, ZoneStore,
+};
+pub use self::tree::{ZoneExists, ZoneTree};
+pub use self::types::{
     Rrset, SharedRr, SharedRrset, StoredDname, StoredRecord,
 };
-pub use self::set::{ZoneExists, ZoneSet};
-pub use self::versioned::Version;
-pub use self::write::WriteableZone;
-pub use self::zone::{VersionMarker, Zone, ZoneData, ZoneVersions};
-
-mod builder;
-mod nodes;
-mod read;
-mod rrset;
-mod set;
-mod versioned;
-mod write;
-mod zone;
+pub use self::walk::WalkOp;
+pub use self::zone::Zone;
